[0.3.x] updated: test
git-svn-id: https://kdlibcpp.svn.codeplex.com/svn@83868 9b283d60-5439-405e-af05-b73fd8c4d996

+++ kdlib/include/kdlib/dbgengine.h
@@ -57,7 +57,7 @@ void setSymbolPath(const std::wstring &symPath);
 void appendSymbolPath(const std::wstring &symPath);
 
 std::wstring getSourceFile( MEMOFFSET_64 offset = 0);
-void getSourceLine( std::wstring &fileName, unsigned long &lineno, int &displacement, MEMOFFSET_64 offset = 0 );
+void getSourceLine( std::wstring &fileName, unsigned long &lineno, long &displacement, MEMOFFSET_64 offset = 0 );
 
 ///////////////////////////////////////////////////////////////////////////////
